azure_app/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
langchain==0.1.9
langchain-openai==0.0.8
requests==2.31.0
supabase==2.3.4
pydantic==2.6.1 